Separate auth fixture from base page
Having both auth and startup in the same fixture leads to issues where it's hard to control for a given test whether it is authenticated or not. It is possible to control this behavior at the test suite granularity, but that's not really convenient. Also, the Playright typing system doesn't make it possible/easy to extend the Page class, which means the resulting fixture only has the base page properties, and that doesn't allow higher-level calls like "login" to go through all the login steps.