Use Lerna to run build
Building based on workspaces alone (i.e. `npm run build --workspaces`) doesn't account for dependencies between workspaces, resulting in an arbitrary build order being based on the globbing pattern/order in which workspaces are defined. Instead, this commit uses Lerna to run the build, and relies on Nx to figure out the build order based on the dependency graph.